Fixed major bug with guerilla turn rules. Print the winner at the end.
As a guerilla player, it's possible to play your first piece without having
a place to play your second piece. The game checks to make sure there are
any valid placements for the guerilla team; if so, it moves on to the second
placement, otherwise it starts the counter-insurgents' turn.

The game now prints the name of the winning team alongside the game over text.