Retarget the project to Android 2.3. Fix orientation bullshit.
By default, Android's Activity subclasses will be destroyed (along with all
of their state) and recreated whenever a) the keyboard comes up, or b) the
phone is rotated. To horrifyingly hack our way around this, we have to
create a singleton that holds all of our state (in this case, a subclass of
the Application class which we know is created when the app is started).

Removed accidental state from the view. The model now has a concept of the
current player.